# internal/adapter/docsui/openapi/tl.yaml
openapi: 3.0.3
info:
  title: ANS Transparency Log API
  version: "2.0.0"
  description: |
    OpenAPI contract for the ANS Transparency Log (`ans-tl`). Paths
    match the reference TL's swagger byte-for-byte — `/v1/agents/*`,
    `/v1/log/*`, `/v1/internal/*`, C2SP tlog-tiles at root
    (`/checkpoint`, `/root-keys`, `/tile/*`), and
    `/internal/v1/producer-keys/*` for admin CRUD.

    Behavioral parity with the reference TL:

      - Event envelope is the V2 shape — nested `{payload:{logId,
        producer:{event, keyId, signature}}, schemaVersion:"V2",
        signature, status}`. V2 evolves the reference V1 with
        unified `identityCerts[]` / `serverCerts[]` arrays, typed
        `dnsRecordsProvisioned[]` records, and `SHA256:`-prefixed
        fingerprints. V0 + V1 schemas are served for historical
        verifier backwards compatibility (never emitted).
      - Receipts are SCITT COSE_Sign1 (RFC 8152 tag 18) with VDS 1
        (RFC 9162_SHA256) and the VDP inclusion proof in the
        unprotected header (label 396). See
        `internal/tl/receipt/constants.go` for the concrete labels.
      - `/root-keys` emits the sumdb-note verification format
        byte-for-byte compatible with the reference.
      - `/internal/v1/producer-keys` admin routes match the reference
        swagger.yaml snake_case DTOs; the only ans-only extension is
        `GET /internal/v1/producer-keys/{key_id}` (reference doesn't
        expose single-record lookup).
servers:
...

...
security:
  - bearerAuth: []
paths:
...
  /v1/internal/agents/event:
    post:
      tags: [Transparency Log]
      summary: Append a signed event to the log
      description: |
        Accepts a JCS-canonicalized producer-event body signed with a
        detached JWS (`X-Signature` header). Successful appends return
        the Tessera-assigned leaf index and hash; duplicate submissions
        (same SHA-256(inner event JCS)) return 200 with the original
        leaf so outbox retries are idempotent.
      operationId: appendEvent
      parameters:
        - name: X-Signature
          in: header
          required: true
          description: |
            Detached JWS compact serialization over the raw request
            body. Header shape:
            `{alg:ES256, kid:<producer-keyId>, typ:JWT, timestamp:<unix-seconds>, raid:<raId>}`.
            The `raid` value must match the `raId` in the body.
          schema:
            type: string
      requestBody:
        required: true
        description: JCS-canonicalized inner producer event (≤ 256 KiB).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerEvent'
      responses:
        '200':
          description: |
            Event logged. On a duplicate submission (identical
            inner-event JCS hash) the response carries
            `duplicate: true` + the pre-existing leaf so retries are
            idempotent. Matches the reference TL's status-code and
            body shape.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppendResponse'
        '413':
          description: Request body exceeds 256 KiB.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: |
            Body fails validation. Codes include
            `NO_PRODUCER_SIGNATURE`, `INVALID_SIGNATURE_HEADER`,
            `NOT_FOUND_PRODUCER_KEY`, `INVALID_PRODUCER_KEY`,
            `MISMATCH_SIGNATURE`, `RAID_MISMATCH`, `INVALID_EVENT`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  # ──────────────────────────────────────────────────────────────
  # Transparency Log — verifier reads (public under public-read)
  # ──────────────────────────────────────────────────────────────
  /v1/agents/{agentId}:
    get:
      tags: [Transparency Log]
      summary: Current badge for an agent
      description: |
        Returns the reference-shaped TransparencyLog envelope for the
        agent's most recent event — including the full inclusion
        proof so clients can verify without a separate tile walk.
      operationId: getBadge
      parameters:
        - $ref: '#/components/parameters/AgentIdPath'
      responses:
        '200':
          description: Badge retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyLog'
        '404':
          description: No events for this agent.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/agents/{agentId}/audit:
    get:
      tags: [Transparency Log]
      summary: Paginated audit history for an agent
      operationId: getAudit
      parameters:
        - $ref: '#/components/parameters/AgentIdPath'
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        '200':
          description: Audit records, newest-first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyLogAudit'
  /v1/agents/{agentId}/receipt:
    get:
      tags: [Transparency Log]
      summary: SCITT COSE_Sign1 receipt for the agent's latest event
      description: |
        Returns the CBOR-encoded COSE_Sign1 receipt (RFC 8152 tag 18)
        containing an RFC 9162 inclusion proof in the VDP header
        (label 396). The `kid` in the protected header is a 4-byte
        SHA-256 prefix of the signing SPKI-DER — cross-reference it
        with `/root-keys` to identify the verifier key.
        Response body is binary CBOR; do not attempt to parse as JSON.
      operationId: getReceipt
      parameters:
        - $ref: '#/components/parameters/AgentIdPath'
      responses:
        '200':
          description: Receipt CBOR bytes.
          content:
            application/scitt-receipt+cose:
              schema:
                type: string
                format: binary
        '404':
          description: No events for this agent.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: |
            The leaf has been appended but no signed checkpoint yet
            covers it. Retry after the `Retry-After` delay.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/agents/{agentId}/status-token:
    get:
      tags: [Transparency Log]
      summary: Signed status token for offline state verification
      description: |
        Returns a short-lived COSE_Sign1 status token (CBOR tag 18)
        asserting the agent's current lifecycle state plus enough
        identity data (cert fingerprints, metadata hashes) that
        verifiers can check a connection without hitting the badge
        endpoint. AHPs fetch this every ~30 minutes and staple it
        alongside the receipt in the Agent Card — analogous to OCSP
        stapling in TLS.

        Content-Type: `application/ans-status-token+cbor`.
        Do not attempt to parse as JSON.
      operationId: getStatusToken
      parameters:
        - $ref: '#/components/parameters/AgentIdPath'
      responses:
        '200':
          description: Binary CBOR status token.
          content:
            application/ans-status-token+cbor:
              schema:
                type: string
                format: binary
        '404':
          description: No events for this agent.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '410':
          description: |
            Agent is in a terminal state (EXPIRED or REVOKED). Status
            tokens are not issued for terminal-state agents; stop
            polling and fall back to the badge for final status.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '501':
          description: |
            Status tokens are not enabled on this deployment. Clients
            should disable status-token polling for this TL.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  # ──────────────────────────────────────────────────────────────
  # Log static artefacts + verifier keys (public under public-read)
  # ──────────────────────────────────────────────────────────────
...
  /v1/log/checkpoint/history:
    get:
      tags: [Tlog Tiles]
      summary: Paginated checkpoint history
      description: |
        Returns checkpoints stored by the TL with optional
        tree-size and time filters. Reference default is 10 per
        page, DESC by tree_size; maximum 100 per page. Matches
        reference swagger §342 byte-for-byte.
      operationId: getCheckpointHistory
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
        - name: fromSize
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 1 }
        - name: toSize
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 1 }
        - name: since
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: order
          in: query
          required: false
          schema: { type: string, enum: [ASC, DESC], default: DESC }
      responses:
        '200':
          description: Paginated list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointHistoryResponse'
        '422':
          description: Invalid query parameters.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/log/schema/{version}:
    get:
      tags: [Tlog Tiles]
      summary: TL event JSON schema for a given version
      description: |
        Returns the JSON schema for the requested TL event version.
        V0 and V1 bodies are the production schemas fetched from
        `transparency.ans.godaddy.com/v1/log/schema/{V0,V1}`,
        served here for historical-verifier backwards compatibility.
        V2 is the shape *this* server emits on the wire (evolution
        of V1 with unified cert arrays + typed DNS records). Matches
        reference swagger §424.
      operationId: getSchemaByVersion
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
            enum: [V0, V1, V2]
      responses:
        '200':
          description: Embedded JSON schema document.
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Unknown schema version.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static API key or OIDC bearer token. Admin routes
        (`/internal/v1/producer-keys/*`) also require the caller to
        be an admin — static keys are always admin in the
        quickstart; OIDC tokens inherit admin-ness from the
        configured admin-groups list. Verifier GETs are anonymous
        when `cfg.auth.public-read: true`.
  parameters:
    AgentIdPath:
      name: agentId
      in: path
      required: true
      description: Immutable agent UUID.
      schema:
        type: string
        format: uuid
  schemas:
...
    ProducerEvent:
      type: object
      description: |
        The canonical inner-event shape an RA POSTs. Must be
        JCS-canonicalized when signed; the server validates the
        signature over the raw request body and re-canonicalizes.
      required: [ansId, ansName, eventType, timestamp, raId, agent]
      properties:
        ansId:      { type: string, format: uuid }
        ansName:    { type: string, example: "ans://v1.0.0.agent.example.com" }
        eventType:
          type: string
          enum: [AGENT_REGISTRATION, DOMAIN_VALIDATION, AGENT_ACTIVE, AGENT_REVOCATION]
        raId:       { type: string, example: "ans-ra-local" }
        issuedAt:   { type: string, format: date-time }
        timestamp:  { type: string, format: date-time }
        agent:
          type: object
          required: [host, name, version]
          properties:
            host:    { type: string }
            name:    { type: string }
            version: { type: string }
        attestations:
          type: object
          description: Event-type specific attestation payload.
    AppendResponse:
      type: object
      description: |
        `logId`, `message`, and `success` match the reference TL's
        AgentEventResponse byte-for-byte. `leafIndex`, `leafHashHex`,
        `duplicate`, and `treeSize` are ans-specific additions that
        expose the Merkle-tree position to callers who want to fetch a
        receipt without a second round-trip.
      required: [leafIndex, leafHashHex, duplicate, treeSize]
      properties:
        logId:       { type: string, format: uuid, description: "UUIDv7 assigned on first append; echoed on duplicate retry." }
        message:     { type: string, description: "Human-readable status message." }
        success:     { type: boolean, description: "Always true for 200 responses." }
        leafIndex:   { type: integer, format: int64 }
        leafHashHex: { type: string, description: "RFC 6962 leaf hash (hex)" }
        duplicate:   { type: boolean }
        treeSize:    { type: integer, format: int64 }

    # ── Badge + audit ─────────────────────────────────────────
...